Add files via upload
FullModelV8.txt is the most successful model I created over the course of my internship. This file was interpreted by SIMION as a GEM file and generated all of the geometric components. 

MagnetFM3.txt is a similar file which generated the magnetic poles that acted as the solenoid in the bending magnet. Later on when I created the electrostatic deflector I didn't use this magnet anymore.